% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_FUNCTION.R
\name{Compute_IE}
\alias{Compute_IE}
\title{Compute Information Energy using KNN}
\usage{
Compute_IE(data_x, data_y, K, mode = "cpp", precision = "float")
}
\arguments{
\item{data_x}{Numeric vector representing the first data set.}

\item{data_y}{Numeric vector representing the second data set.}

\item{K}{Integer specifying the number of neighbors.}

\item{mode}{Character string specifying the mode ("cpp", "cuda", or "openCL").
Defaults to "cpp".}

\item{precision}{Character string specifying the precision ("float" or "double").
Defaults to "float".}
}
\value{
The result returned by the native C++ function.
}
\description{
This function calls the exported C++ function `KNN` directly.
}
